# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OSL Old Main Server API
  version: 1.0.0
  description: REST API for the main e-commerce server (monolith)
servers:
  - url: http://localhost:4000
    description: Local
...
paths:
...
  /banners/{id}:
    get:
      tags: [Gallery]
      summary: Get banner by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  # Gallery (public)
...
  /reviews/product/{productId}:
    get:
      tags: [Reviews]
      summary: Get product reviews (public)
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  # General reviews (mixed)
  /reviews/{reviewId}/likeCount:
    get:
      tags: [Reviews]
      summary: Like count
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /reviews:
    post:
      tags: [Reviews]
      summary: Create review
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
  /reviews/{reviewId}:
    put:
      tags: [Reviews]
      summary: Update review
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Updated }
    delete:
      tags: [Reviews]
      summary: Delete review
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
  /reviews/{reviewId}/like:
    post:
      tags: [Reviews]
      summary: Like review
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /reviews/{reviewId}/unlike:
    post:
      tags: [Reviews]
      summary: Unlike review
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  # Payments
...
  /users/reviews/product/{product}:
    get:
      tags: [Reviews]
      summary: Get my review for a product
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: product
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /users/reviews:
    post:
      tags: [Reviews]
      summary: Create my review
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
  /users/reviews/{id}:
    put:
      tags: [Reviews]
      summary: Update my review
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    delete:
      tags: [Reviews]
      summary: Delete my review
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }

  # Admin endpoints (summarized key routes)
...
  /admin/banners/{id}:
    get:
      tags: [Admin - Banners]
      summary: Get banner by id (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    put:
      tags: [Admin - Banners]
      summary: Update banner (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
                imageUrl: { type: string, format: uri }
                pageLink: { type: string }
                active: { type: boolean }
                category: { type: string, enum: [A, B, C, D, E] }
      responses:
        '200': { description: OK }
    delete:
      tags: [Admin - Banners]
      summary: Delete banner (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...