# src/routes/general/review.ts
import express from 'express';
import { authenticateUser } from '../../middleware/auth';
import ReviewController from '../../controller/reviewController';
import ReviewValidator from '../../validators/ReviewValidator';
import ReviewService from '@/services/reviewService';

const router = express.Router();

// Public list endpoint: reviews for product
router.get('/product/:productId', async (req, res) => {
  try {
    const { productId } = req.params;
    const page = Number(req.query.page || 1);
    const { data, message, code } = await ReviewService.allReviews(productId, page);
    return res.status(code).json({ message, data });
  } catch (error) {
    console.error('Error listing product reviews:', error);
    return res.status(500).json({ error: 'Internal server error' });
  }
});

// Public routes (no auth required)
router.get('/:reviewId/likeCount', ...ReviewValidator.validateReviewId, ReviewController.getLikeCount);

// Protected routes (auth required)
router.post('/', authenticateUser, ...ReviewValidator.validateCreateReview, ReviewController.createReview);
router.put('/:reviewId', authenticateUser, ...ReviewValidator.validateUpdateReview, ReviewController.updateReview);
router.delete('/:reviewId', authenticateUser, ...ReviewValidator.validateReviewId, ReviewController.deleteReview);

router.post('/:reviewId/like', authenticateUser, ...ReviewValidator.validateReviewId, ReviewController.likeReview);
router.post('/:reviewId/unlike', authenticateUser, ...ReviewValidator.validateReviewId, ReviewController.unlikeReview);
router.get(
  '/:reviewId/isLikedByUser',
  authenticateUser,
  ...ReviewValidator.validateReviewId,
  ReviewController.isLikedByUser
);

export default router;